Se creo el controller
