+++ src/controllers/DocumentController.php
@@ -31,15 +31,7 @@ class DocumentController
     }
 
 
-    public function showCommission($role, $userID)
-    {
-        $documents = $this->documentModel->getAllDocuments($role, $userID);
-        $diasEconomicos = $this->documentModel->countDiasEconomicos($userID)['diasEconomicos'];
-        $diaCumple = $this->documentModel->countDiaCumple($userID)['diaCumple'];
-        $reportesIncidencia = $this->documentModel->countReportesIncidencia($userID)['reportesIncidencia'];
-        require VIEW_PATH . 'document/commission_list.php';
 
-    }
 
     public function downloadDocument($docID)
     {
